Move region metadata classes to new metaclass.py module
